:recycle: Relocate tests
