+++ src/server/handler/delete.rs
@@ -16,3 +16,19 @@ pub async fn delete(
     let value = state.storage.delete(&resource, &id)?;
     Ok((StatusCode::OK, Json(value)))
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::server::handler::delete::delete;
+    use axum::extract::{Path, State};
+
+    #[tokio::test]
+    async fn test_delete() {
+        let state = crate::server::handler::tests::init_state();
+        let path: Path<(String, String)> = Path((
+            "posts".to_string(),
+            "01J7BAKH37HPG116ZRRFKHBDGB".to_string(),
+        ));
+        assert!(delete(path, State(state)).await.is_ok());
+    }
+}
